Update patch set 4
Patch Set 4:

Thanks for your feedback. We can wait until you're back, there's no rush.

The approach that I came up with and developed in this CL is to apply a review-only CSP policy and then listen and capture any resulting violation reports. The alternative you propose is to simply enforce a CSP policy in the expectation that tests should fail if violations occur -- and I think this is a good option also.

The advantage of listening for violation reports is that we can test for a wider variety of issues. For example, in the previously-fixed Anchor class, the issue was that the the href was set to "javascript:;". It's not really possible to detect this or test a fix with the second approach (the href is a no-op).

It's not the intent of this CL to establish new APIs. The only APIs that are intended to be visible are the gwt.strictCspTestingEnabled property and the CspViolation exception class. Do you consider the ExpectedFailure and related classes new APIs? Therese are slightly modified existing classes are were moved from the test/ directory to allow the CspTest test to build. Maybe we can refactor to expose fewer of these APIs?

The checkpoints were needed to attribute CSP violation reports correctly. Without them, it's hard to know which test generated which CSP violation.

We can just start with the second approach for this CL (e.g. remove the report-capturing logic and change the CSP mode to enforcing). Let me know if that sounds good.

Patch-set: 4
Tag: mailMessageId=<CAJGyu7WbtbuKtfeRxd2FyJnXFyfT8Cee7qnCFzjxpfBBijSw4A@mail.gmail.com>