Update patch set 5
Patch Set 5:

> Let's start with the second approach.
 > 
 > When CSP violation happens, isn't that reported somehow like via
 > window.onerror?

I've refactored to this approach. Please take a look.

I tried moving the config property to JUnit.xml, however I encountered an issue and reverted it back to how it was. Files like UserTest.gwt.xml and DomTest.gwt.xml, where we want to be able to set the property, do not inherit from JUnit.xml, where the property is defined.

We can detect CSP violations in JavaScript by listening for the SecurityPolicyViolationEvent, but this works only at Chrome, at least for the time being. We can use it, and it would be a little bit better because it would simplify the code a little bit and we wouldn't have to worry about HTTP requests being sent in the correct order. We'd still need to use checkpoints though to correctly attribute violations because they are generated asynchronously.

Patch-set: 5
Reviewer: Gerrit User 10840 <10840@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>