Update patch set 5
Patch Set 5:

The discussion separated on couple of threads so I will reply here.

w.r.t if you want performance, use char[]:
That is not really practical for two reasons;
1. Getting char[] out String is even costlier than new version of charAt.
2. This is a very fundamental API and the code is already out there which uses this extensively with the assumption of the cost.

w.r.t Performance doesn't seem to be a factor in the InternalPreconditions decision making:
Not completely true, otherwise we would only have checkCriticals and wouldn't need to make a decision. I normally lean towards debuggability and tried to make performance out of equation but at the end this is always a judgement call especially for very critical APIs.


To be honest, I was actually going to say no to any checking here, because the behavior we have here is very similar to Array (Array vs String index access) and we have similar performance concerns. I am OK with it just because the debugging is difficult and InternalPrecondition's not-critical checks is a good tradeoff between performance and debuggability.

If we have a check here we will lose our edge in any String manipulation code and push people to native code if they care about performance which would be really bad.

Patch-set: 5
Reviewer: Gerrit User 5110 <5110@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>