Update patch set 5
Patch Set 5:

Sorry if I was unclear - I don't mean to suggest that performance doesn't matter, but only that InternalPrecondition doesn't weigh it when considering "how do we make sure that Java semantics are respected". 

Can we all agree that a BOUNDS-level check is good middle ground, allowing users to turn it on or off? I'd love for this to be more tuneable logging, but this is the system we have today, with only three possible settings based on what users might want to have checked, and what parts of the JRE emulation they are expecting to be inaccurate.

Alternatively (or additionally, long term), support disabling critical checks - some odd things seem critical given this NaN issue, things that probably shouldn't be critical at all.

FWIW, I am not sure if TeaVM beats us here in terms of performance (as they do in many areas), but they make a point of respecting the expected Java semantics, and generally seem to get better performance through the use of typedarrays. That of course comes at a cost when interop'ing with JavaScript, since strings, arrays of numbers need to be converted in each direction.

Patch-set: 5
Reviewer: Gerrit User 5130 <5130@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>