Update patch set 7
Patch Set 7:

OK, so, back to the whiteboard.

Assuming the original reporter leaved in St Petersbourg at the time he reported the issue (he since moved to Australia, from his phone number in his resume; and I believe the issue would be the same –actually worse– seen from there), that put him in Europe/Moscow time.

The date used for testing is 1997-10-26T06:41:10Z. The target timezone (America/Creston) is UTC-7, so the date there is 1997-10-25T23:41:10-07:00. In 1997, Russia observed DST, with 1997-10-25 in DST (UTC+4), and 1997-10-26 not (UTC+3; switching at 2am/3am); note that in 2013, Russia was in "permanent DST" (UTC+4), since 2011 and until 2014 when switched to UTC+3 with no DST (but GWT might actually not be aware of that last change, as reported today!). The difference between OS timezone and target timezone is then 10 hours.

Adjusting the date (1997-10-26T09:41:10+03:00) by 10 hours then gives us, because of DST, 1997-10-26T00:41:10+04:00. keepDate is thus wrong (26, not 25), and keepTime is adjusted "by 1 hour" because the date changed timezone, so we get it right (23:41:10).

For reference, the way keepTime is adjusted is by adding/substracting a full day. Here, date.getTimezoneOffset() is -180, and timeZone.getOffset(date) is 420 (signs are reversed compared to their text representation), so 'diff' is -600*60000. Because it's negative, keepTime is adjusted to "the next day", which puts it to 1997-10-26T23:41:10+03:00 such that we cross the DST line again.

So, the problem is crossing the DST limit, when that limit is *not* at midnight, so you might be left on the same date. In this case, because keepTime actually didn't changed date (was 1997-10-26T00:41:10+04:00, adjusted to 1997-10-26T23:41:00+03:00) when we "adjusted it", we could update keepDate accordingly. Let's see if that would be the appropriate fix.

The tricky part, as explained in the comment above, is when switching is DST: there's one hour that simply does not exist.

Let's change to the date to be formatted to something like 1997-03-30T09:41:10Z, aka 1997-03-30T13:41:10+04:00. Adjust it by 11 hours (because were now in DST, but America/Creston is fixed at UTC-7), it gives you 1997-03-30T01:41:10+03:00. We crossed the DST line again, so we need to adjust keepTime to make it 02:41:10, but that time simply does not exist on 1997-03-30 in Europe/Moscow. keepTime is thus adjusted to 1997-03-31T02:41:10+04:00.

What about our proposed fix? Well, here, keepTime changed date (went from 1997-03-30 to 1997-03-31), so we'd keep keepDate the same; and it would be right!

So, how about that fix?

 if (/* keepTime crossed DST, i.e. needs to be adjusted */) {
   int day = keepTime.getDate();
   // adjust keepTime as usual
   if (keepTime.getDate() != day) {
     // adjust keepDate in "reverse" direction compared to keepTime
     // (when we added to keepTime, we need to substract to keepDate)
   }
}

The way to adjust keepDate is "to be defined" though. Moving it by 24h might possibly not be enough, so maybe first reset its hour to 0 and then move it by 24h?

I'll let you think about it, I need to go now :)

Patch-set: 7