Fix date formatting around daylight time transitions
The date can be off by one full day in two cases:

When switching out of DST, the day is longer than 24 hours;
when re-adjusting the time, it can then end up in the same
day. When that's the case, it means the date needs to be
pushed one day in the other direction.
E.g. 2006-10-29T00:30:00-07:00 (adjusted date) + 24h pushes
it to 2006-10-29T23:30:00-08:00 (adjusted time). In this
case, the correct date would be 2006-10-28, not 2006-10-29.

When switching to DST, the day is shorter than 24 hours;
if the adjusted time ends up around midnight, it can end
up on the wrong day. In this case, adjusting the time
back into the proper timezone will push it on an entirely
different day. When that's the case, keepDate can then
use the original date instead.
E.g. 2006-04-02T03:30:00-07:00, when adjusted for UTC-10:00,
ends up at 2006-04-01T23:30:00-08:00. The time is then
re-adjusted to 2006-04-03T00:30:00-07:00. In this case,
the correct date would be 2006-04-02, not 2006-04-01 (which
is necessarily the same as the initial date).

Bug: #8259
Bug-Link: https://github.com/gwtproject/gwt/issues/8259
Change-Id: I5ecf7761eed28b088c2085a13af772505b4bdfe9