Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> If all of the leaks (and we're talking about leaks in Jetty, right?)

No, we're talking about leaks due to "pinned" classloaders (i.e. the WebAppClassLoader is referenced by a class in the parent/current classloader so it cannot be garbage-collected).
The leak is then experienced when restarting the webapp (where the WebAppClassLoader is "destroyed" and a new one created). This specifically does NOT apply to the JUnitShell for instance, only to DevMode's embedded web server (this is one reason I don't think we should spend too much time/energy on it); but the message shows up in GWTTestCase because the same code is run (this is the reason I think we should do something anyway).

One would probably have to check whether Jetty does that sort of thing already in the version we're using (it looks like it might have been added in Jetty 7.6.6 [1], now we'd have to see whether it's enabled by default and/or how to enable it). Also, Jetty apparently simply ignores the ClassNotFoundException, without even trying to detect Java 9 vs. older versions [2] (as I was suggesting)

[1] https://github.com/eclipse/jetty.project/commit/0b4d2e6a0fe6b37363ee0caf7f84b420b7c74cc2
[2] https://github.com/eclipse/jetty.project/blob/71d5fcf246034dce91cd642d35cdbaace9806329/jetty-util/src/main/java/org/eclipse/jetty/util/preventers/GCThreadLeakPreventer.java

Patch-set: 1
Reviewer: Gerrit User 5030 <5030@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>