Update patch set 6
Patch Set 6:

I'll change the approach to only avoid ordinalizing the ones that are passes to a method that canbeimplemeted externally. Or are the return of methods that can be referenced externally. Need to add quite a few tests.

Patch-set: 6
Tag: mailMessageId=<CAC7T7gnGYq-nTQHOwoMwg1SQAGQrHQVHWhUAYM3AxRzO8QS1_Q@mail.gmail.com>