Update patch set 1
Patch Set 1:

> It looks like there are multiple instances of MinimalRebuildCache in memory and I guess it 
... 
> At the end of the day I had 15 instances of MinimalRebuildCache in memory each retaining roughly 120 MB => 1.8 GB.

There should be up to 3 live MinimalRebuildCache instances (see MinimalRebuildCacheManager) to speed up when you work with different permutations. There should be only one if you are working with the same permutation all the time. This is all done through codeserver.Recompiler which is what we use here for our SDM integration.

Not sure if the eclipse plugin integration uses this route tough, IIRC it uses the dev mode integration (DevModeBase), which I am not sure it uses the Recompiler at all. A summary reading of the code suggests that the integration in DevModeBase creates these new MinimalRebuildCache instances. But I am not really familiar with this code.

Patch-set: 1
Reviewer: Gerrit User 5105 <5105@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>