Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Those are good news. I made a small change to use SoftReference instead of WeakReference to avoid unnecessary reparse of the class data if there is enough memory. This one should also work fine, could you check? In theory these references will be kept around unless there is memory pressure.
> 
> I'll check it on Monday or Tuesday. 
> 
> I was wondering if CompiledClass.getNameEnvironmentAnswer() could be deleted and call sites implemented differently. The usage in JdtCompiler could probably be replaced with a mapping from String (internalName) -> NameEnvironmentAnswer. I am unsure about the usage in CompilationStateBuilder though as I am not familiar with the code. Maybe the same mapping could be used in CompilationStateBuilder as well.

I think that would probably defeat the whole purpose of keeping the reference which is caching the answer.  My guess is that getEnvironmentAnswer is called over and over and over on the same CompiledClass as part of the compilation. Either there is a reference to an answer in CompiledClass and we have the leak or there is no caching and pay the cost of recomputing it.

Other option is to only cache the classreader object and recreate the answer every time. That might probably work assuming most of the cost if on the class reader part (which is probably the case).

I might try that which seems a better solution overall. Please check if this patch also works and I'll upload the classreader solution tomorrow for you to try.

Patch-set: 3
CC: Gerrit User 5305 <5305@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>