Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> (1 comment)
> 
> > Patch Set 3:
> > 
> > (2 comments)
> > 
> > > Patch Set 1:
> > > 
> > > (4 comments)
> > > 
> > > Wouldn't it be even better to remove support for Firefox <= 4.0? It's been EOL since June 2011, that's 7.5 years ago! (interestingly, Firefox 3.6 EOL was more recent, but still: April 2012 ! and we're talking about EOL here, not "last release"). Fwiw, those EOL dates come from Wikipedia: https://en.wikipedia.org/wiki/Firefox_version_history#Early_releases_through_version_4
> > 
> > While I could not agree more on this comment, I originally did not want to remove any features/support from GWT (knowing there might be still people using those old browsers) and do a simple workaround for those edge cases.
> 
> I understand that, and thank you for taking those precautions. Let's see what other reviewers think.
> 
> If we go that road, we'd want to remove the similar code from PopupImplMozilla too (a grep on 'rv:' gives only those places where we extract Gecko version); or actually, we could remove PopupImplMozilla entirely, and possibly even inline PopupImpl into PopupPanel (and remove the outdated comments about hidden iframes from the IE6 era).

Shouldn't it be OK as in PopupImpl the code is already stored as a static boolean? 

  private static boolean isFF2Mac = isFF2Mac();

So, the road towards removing PopupImpl could be the one to take. 



On the old browsers support topic, if I can help the other reviewers, I have been in contact with some end users running something called Mozilla Prism in the past few years and it is (on the top of my head) recognized as Firefox 2.5. I think they migrated now to some IE10~ish version.

Patch-set: 4
Reviewer: Gerrit User 11267 <11267@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>