Update patch set 4
Patch Set 4: Code-Review-1

Giving this a self -1 for the moment to prevent a +2-and-merge, but I am looking for feedback.

Nothing actually runs during a build, as even if you have Java 9, the build itself doesn't work in Java 9+. Some work will be required for this, I haven't dug too deeply, but these individual tests do pass.

Some parts of this are a bit inconsistent, and I wanted feedback on the patch so far to see if there is a preferred style to use.

(Optional, Collectors, Stream patches are right behind this, followed by Java 10 patches for Collectors and List/Map/Set.)

Patch-set: 4
Reviewer: Gerrit User 5130 <5130@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=-1