Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (3 comments)
> 
> > Patch Set 1:
> > 
> > We might need to use a newer JDT, see:
> > 
> > https://git.eclipse.org/c/jdt/eclipse.jdt.core.git/commit/org.eclipse.jdt.core?id=4c49f905582020b5bbace350773b0eba3dc92fb8
> > 
> > Maybe create a test for Predicate and Function to see if it fails.
> 
> I've added test for Function and it fails.
> https://gwt-review.googlesource.com/c/gwt/+/21540/2/user/test-super/com/google/gwt/dev/jjs/super/com/google/gwt/dev/jjs/test/Java11Test.java#53
> 
> So far, I have it commented out. I think we may to proceed with this patch and then just upgrade JDT when Eclipse 4.11 will be released. Or I may to create a custom build with cherry-picked fix for this bug (JDT milestone may be a bit unstable).

I think I would defer this CL until 4.11 is released. There are a couple of other bugs fixed that seem useful. Also because quite a few people use GWT-HEAD it is better to not support something instead of half support it. Alternatively change the commit message to reflect the current state, or disable SourceLevel 11 for the time being until JDT has been updated again.  Updating JDT to 4.11 should be a blocker issue in the 2.9 milestone then.

Other useful Bugfixes:
- Slow compilation: https://bugs.eclipse.org/bugs/show_bug.cgi?id=541387
- Good code is bad in JDT: https://bugs.eclipse.org/bugs/show_bug.cgi?id=543526
- Type inference bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=543128

Patch-set: 2