Link backing errors together with a cause attribute
We currently only link backingJsObjects for Throwables bidirectionally between
the JS object and the Java Throwable. We then only chain the Throwable caused
bys. However, there are situtation where it would be useful to have the JS
errors linked together, in particular when an uncaught error gets handled. We
could traverse back and forth between the Java Throwable and the backing object,
but that's not always feasible (ex. when running in v8/jsc directly and the
error object gets handled on the native side).

Author: kevinoconnor
Change-Id: I43102d3e9796ac307a4acdfff46263e903106daf