Update style check to be more consistent with Google Java style.
The aim of this patch is to reduce friction and make the styles from
GWT and J2CL the same. In particular J2CL source code is autoformatted
to match the style guide and such code does not conform the the
current GWT style.

Change-Id: I6bb167b6c37ab617e7e7ba8abcb39d7f4bd4d64d
Review-Link: https://gwt-review.googlesource.com/#/c/21980/