Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+2
> 
> Only thought I would have: do you want to include the reformat pass in this same commit, rather than wait and merge it separately? This way we could see the cause/effect in a single review.

Thanks. The current format works is a bit less stringent and works for our current GWT standards and for J2CL. The changes are:
 - does not require lines separating different types of imports but if there are it is fine as long as every group is sorted.
- allows 1 line JavaDocs.
- allows empty braces with no space in between. 

We have the autoformatting integrated with our internal code review system. I'll have to see how to integrate it to gerrit or git. Let's do that as a follow up.

Patch-set: 2