Update patch set 2
Patch Set 2:

It is mangled from gmail sorry... Retrying...

Would something like this be adequate?

Two verbose? How about?

p1 only?
p1 && p4 only?

/**
The current implementation supports setting correctly the function.length for lambdas with up to 10 arguments. If a lambda has more than 10 arguments it would be constructed and called correctly but function.length would be set to 10.


See github issue #9675 for details.


In the original implementation the argument count of the functions was set to 0 disregarding the actual input arguments. The current implementation does not prevent to create lambda with > 10 arguments. It just corrects the argument count (function.length) in the cases <= 10 hoping that it would catch most of API exported by JS libraries.


If the need arises in the future to support argument counts > 10 the required extension of the function is trivial and straightforward.
*/

Patch-set: 2