Update patch set 2
Patch Set 2:

Hi,

First of all, thanks for contributing. 

As I expressed in the related bug I don't think this is the right approach. The right approach IMHO is to do as J2CL is doing, where it passes @JsFunction lambdas directly as JavaScript functions without wrapping them as classes. 

Given that it is bad practice to have different behaviour depending on the number of parameters a function declares, I think this problem should be workaround in the user code as I suggested in the bug.

Patch-set: 2
CC: Gerrit User 5105 <5105@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>