Update patch set 1
Patch Set 1:

(4 comments)

I can see a point to mark literally everything deprecated, but it will be a bit more time consuming, and the type has to be mentioned to create instances or reference static members, return from JSNI methods, etc, so I think at least the developer will get the point.

Agree we want a note in release notes about this, something to the effect of "This release of GWT supports Elemental2's 1.0.0 release (link), so the gwt-elemental jar and its contents have been deprecated. Elemental2 will be regularly updated, and uses JsInterop, so is also compatible with J2CL. The gwt-elemental.jar may be removed in a later release."

Patch-set: 1