Fork WebAppCreator task to use plain java classloader
When ant's classloader references a directory within a jar and is
run from Java 8+, it incorrectly appends a trailing slash to the
java URL instance. For WebAppCreator, this is a breaking change,
and forking this task brings it in line with the other <java> tasks
throughout the project.

With that changed, samples in the gwt release zip correctly include
build files, readmes, and some other contents.

Change-Id: I5ba2a3024e07c7f817716a4c4c428acd745405bf