Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> with JDK 11.0.1 and jdk14.0.1 I got followíng error:
> 
> Caused by: java.nio.file.NotDirectoryException: C:\Java\jdk-13.0.2\lib\modules
> 	at java.base/sun.nio.fs.WindowsDirectoryStream.<init>(WindowsDirectoryStream.java:84)
> 	at java.base/sun.nio.fs.WindowsFileSystemProvider.newDirectoryStream(WindowsFileSystemProvider.java:520)
> 	at java.base/java.nio.file.Files.newDirectoryStream(Files.java:475)
> 	at java.base/java.nio.file.Files.list(Files.java:3762)
> 	at com.sun.tools.javac.file.Locations$BootClassPathLocationHandler.systemClasses(Locations.java:665)
> 	at com.sun.tools.javac.file.Locations$BootClassPathLocationHandler.computePath(Locations.java:621)
> 	at com.sun.tools.javac.file.Locations$BootClassPathLocationHandler.lazy(Locations.java:709)
> 
> This error still exist without your change but should it possible to build with Java 9+

Can you confirm that this is specifically from the WebAppCreator? Or is this somewhere else in the build?

At least the doctool has issues with being invoked on newer JREs, and we'll need to update errorprone as well in order to build (as gwt's ant wiring declares it today), but the tooling should all work in Java11, 14 (and this was confirmed in testing so far, on a variety of platforms), including the webappcreator tool.

Patch-set: 1