Drop support for building or running on Java 7.
This patch does some initial work of adding support for building GWT
on Java >8, but until the javadoc wiring is rebuilt (or conditionally
disabled) a normal build won't be possible.

After Javadoc has been updated or guarded behind a flag, isJava8 can be
false when Java >8 JDK is used to build, and tests for those can be
guarded as makes sense.

The error-prone Java compiler is also temporarily disabled in this
commit, and will be restored and updated to latest in a follow-up patch.

Bug: 9683
Bug-Link: https://github.com/gwtproject/gwt/issues/9683
Change-Id: I5ec65e330409802368b2a60148d7ef1ee97c03d1