Update patch set 1
Patch Set 1:

> At some point the continuous integration server would need to be hosted somewhere else so that the community can maintain it easily.

We can run the build externally, but it will need to be deployed to sonatype-snapshots from Google-managed infrastructure, unless it is possible to grant someone outside Google the rights to deploy to com.google.gwt namespaces in sonatype/mavencentral. I'm not sure it makes sense to duplicate parts of the build, and skip tests on just the google version?

Patch-set: 1