Introduce a new setAt() method for EnumMap.
J2CL uses the GWT emul sources, and the same sources also transpile to WASM
(unless the entire implementation must be different using super-wasm sources,
which is generally not true for collections classes). For these classes, array
elements should not be read or assigned out of bounds, which will result in a
WASM exception (for WASM-GC arrays). ArrayHelper will have a super-wasm
implementation, which will resize the backing array if needed before
proceeding.

Like the original EnumMap helper method, this returns the original value at the
index (or else null), which can be done efficiently in GWT without bounds
checks.

PiperOrigin-RevId: 378638984
Change-Id: Ia19caef0576f8064a23b7cc73b20eea445e18b60