Fix buggy tests discovered while upgrading HtmlUnit
The selection script test incorrectly let the default html page load for
a .nocache.js file, but this did not fail in older htmlunit. Newer
htmlunit validates the content-type and fails if html is loaded instead
of js content.

The Jsonp test was previously excluded in htmlunit because there was no
counter in use, and the test tried to call finishTest() twice. This is a
potentially a test bug in other browsers as well, since the second call
could fail or never finish, and the test would not fail.

Change-Id: Ib8de43d4a6911ceaa4354bf6650147bb76664b32