Enable GitHub Actions for pushes, nightly builds, and PRs (#9762)
This introduces two builds, intended to fairly closely reflect the
builds that used to be run on the build.gwtproject.org Jenkins instance.

Nightly builds run most tests, and provide both a simple visualization
of the test summary as well as access to download all test reports, and
in the case of successful builds, the sdk zip can be downloaded. Maven
artifacts are not yet available. This full build can also be run by
pushing to main or release branches, and can be run by creating branches
that end with "-test", which may be helpful for forks to run their own
actions.

On each Pull Request, only checkstyle and apicheck will be run, in
addition to verifying that the SDK and samples can be built, but the ~2
hours of tests are not run. Any checkstyle errors that overlap with the
git diff will be posted as comments or annotations to the diff.

Partial #9756