Deploy nightly builds to sonatype, using github secrets (#9763)
Configures nightly scheduled builds to push to sonatype snapshots, as
used to happen with the Google-managed Jenkins job. GitHub Actions
for the account is assumed to have SONATYPE_USERNAME and
SONATYPE_PASSWORD secrets configured. To avoid forks attempting to
deploy to org.gwtproject from their own actions, this is also set to
only run on the gwtproject organization's repository.

Partial #9756