Add Jakarta servlet support, only for gwt-servlet
This is a draft patch to add support for Jakarta. It is assumed to only
be used for modular projects, where client and server do not live in the
same classpath. Among other things, this means that validation does not
need to be updated, and old javax.servlet references can be left in
gwt-dev and gwt-user.

This does not correct the sources for deployment to maven, nor does it
fix javadoc.

This does not update the requestfactory-server jar.

Draft for #9727