Correct Java versions numbers to support -sourceLevel auto
Starting in Java 9, the value of the system property
java.specification.version has been an integer, with no "1." prefix. As
a result, using -sourceLevel auto with any version 9 and above has
automatically used the latest version of Java that GWT supports, rather
than an actual matching version.

This doesn't matter a lot today for Java 9 and 10, but with coming Java
17 support while 8 and 11 are still allowed, it may make a difference.

Partial fix for #9813