Update installScriptEarlyDownload.js to remove old workaround (#9796)
The eval behavior in GWT is no longer needed since that chrome bug
(#90707) was fixed. Removing that behavior improves the performance
of bootstrapping. It also coincidently fixes the referenced CSP issue.

Considered unifying the logic between browsers, but Safari took a bit
of a perf hit when trying to insert a single large text.

Fixes #9725