Java9+ support for doctool (#9780)
Update doctool to run on Java 11, thus allowing either Java 8 or 11 to
build GWT itself. Because the new javadoc types are not available in
Java 8, the ant wiring has been configured to just skip docs when
building with Java 8. CI is updated to build with both 8 and 11, but
since Java 8 cannot produce documentation, only the Java 11 build is
published.

Fixes #9683