Deprecate and remove references to RPCServletUtils.CHARSET_UTF8
This was necessary pre-Java7 to avoid blocking on
java.nio.charset.Charset operations, though this not only appears to be
resolved in the JRE, but also made unnecessary for our use case through
the use of java.nio.charset.StandardCharsets.

Additionally, this removes a dependency on javax.servlet specific code
from generic code that could be shared with a jakarta.servlet
implementation.

See #6397
Partial #9727