Run build, tests in Java 17 (#9830)
With Javadoc now able to run on Java 17, we can make Java 17 the default
going forward for builds. This patch enables running tests in 17, and
fixes a few gotchas discovered - an old asm jar was being used for
tests, and when we updated to the latest asm for GWT 2.10, we didn't
update ClassVisitors to use the newer version.