Log deprecation warning when using built-in JettyLauncher
This patch has two approaches to check if the JettyLauncher is being
used, it checks for any configured filters/servlets (other than the
defaults provided by Jetty) in order to log a warning on startup, and
it checks for non-404 responses to catch other resources loaded after
startup.

No warning is logged for tests.

Fixes #9863