Log deprecation warning when using built-in JettyLauncher (#9866)
This patch has two approaches to check if the JettyLauncher is being
used, it checks for any configured filters/servlets (other than the
defaults provided by Jetty) in order to log a warning on startup, and it
checks for non-404 responses to catch other resources loaded after
startup.

No warning is logged for tests which use servlets.

A new ServletContainerLauncher implementation is added, to be passed as
`-server com.google.gwt.dev.shell.StaticResourceServer` to DevMode. 
This will serve only static contents instead of hosting the servlet-
based web application. In the future, this may be the default in place
of JettyLauncher.

Fixes #9863