Remove unexpected sources from gwt-user by making valid servlet sources
Removes an old hack which uses gwt-user's sources as if they were the
correct sources for gwt-servlet by creating a real build artifact that
contains the sources which correspond to the classes found in
gwt-servlet and gwt-servlet-jakarta. This lets us avoid using gwt-user's
own sources as if they were the sources for other jars, which in turn
allows us to only package gwt-user sources in gwt-user.jar.

Fixes #9886