Remove all user agent checks for MSIE or documentMode
This removes all checks of MSIE in the user agent string, and all checks
of documentMode to verify which IE version is being used, as these
browsers are long since unsupported.

Most of these changes are specific to tests or devmode, but there is
also a change to simplify Throwable slightly (allowing it to work in
node.js and other non-browser environments), and simplification of
startup user agent checks.