Allow lambdas in this()/super() to close over outer types (#10078)
Prior to JDT 3.32, passing a lambda that closes over outer types to
this()/super() constructors would result in a compiler error. As of
around 3.32 this is now allowed, but the closed-over instance is passed
as an item in a "synthetic outer local variable" collection. Each
instance in that collection has a field representing the "actual outer
local variable", and in this case, that field is null.

Added tests to confirm that the generated code seems consistent, and
that running such a lambda affects the expected instance.

Fixes #10065