Change default SCL instance, and allow selecting from service loader
The default is now to ask a service loader for a
ServletContainerLauncher if there is a single entry available, and
failing that to use the static server - no servlets will be available.

ServletContainerLaunchers registered as a provider for the service
loader must provide a name that matches a specific regex. Those names
may now also be used when specifying a particular server name rather
than the fully qualified class name.

The JettyLauncher will be kept for at least another version to ensure
that projects can easily migrate to newer versions.

This patch also moves `setBaseRequestLogLevel` to
ServletContainerLauncher so other implementations can use it to log
appropriately based on how they were started.

Fixes #10057