Clean up file hash usage (#10102)
GWT currently uses a mix of md5 and sha1, though all use cases are
non-cryptographic (and both hash functions are discouraged, and have
been since before they were added to GWT). This patch proposes replacing
them with Guava's murmur3 128 implementation, as it is slightly faster
than md5 (and much faster than sha1), and can still be used to produce
the same output format (16 upper-case hex digits).

As part of this refactor, some structured content now has size or field
delimiters hashed to avoid collisions, and some streamed content is
hashed as it is read to avoid holding the entire payload in memory at
once. The Util method dealt with the size delimiters in some cases, but
not uniformly, but also ignored exceptions - this patch also cleans up
those errors to correctly throw with meaningful errors.

Fixes #10090