Improving accessability - added aria presentation role for layout tables (#10108)
According to accessibility rules layout tables exist merely to position
content visually and should not be used in HTML5.
Screen readers may interpret them as data tables (i.e., announcing
column and row numbers) etc). To change that in GWT would be a bigger
task. So as a minor improvement those tables should be marked as layout
tables. This can be done by assigning the role="presentation" to ensure
it is not identified as a table to screen reader users.

---------

Co-authored-by: Colin Alworth <colin@vertispan.com>
Co-authored-by: Zbynek Konecny <zbynek@geogebra.org>